feat: add card modules and tests
