+++ Fortschritts.txt
@@ -1,1 +1,1 @@
-Fortschritt: 20%
+Fortschritt: 30%
