+++ app.js
@@ -85,6 +85,11 @@
     const st = document.getElementById('status');
     if (st) st.textContent = 'Fehler: ' + e.message;
   });
+  // Fange unbehandelte Promise-Ablehnungen ab
+  window.addEventListener('unhandledrejection', e => {
+    const st = document.getElementById('status');
+    if (st) st.textContent = 'Fehler: ' + e.reason;
+  });
 
   /* Zustand laden oder initialisieren */
   let state = loadState() || initState(today.getFullYear());
@@ -214,18 +219,6 @@
     rootStyle.setProperty('--btn-on-primary', lum > 0.6 ? '#000000' : '#ffffff');
     state.palette = name;
     safeSet(PALETTE_KEY, name);
-  }
-  function hexToRgb(hex) {
-    const h = hex.replace('#','');
-    const n = parseInt(h,16);
-    return { r:(n>>16)&255, g:(n>>8)&255, b:n&255 };
-  }
-  // Konvertiert eine Hex‑Farbe in einen rgba‑String mit definierter
-  // Transparenz. Wird genutzt, um Monatsüberschriften mit einer
-  // leicht transparenten Version der Akzentfarbe zu hinterlegen.
-  function hexToRgba(hex, alpha) {
-    const { r, g, b } = hexToRgb(hex);
-    return `rgba(${r},${g},${b},${alpha})`;
   }
   function luminance(r,g,b) {
     r/=255; g/=255; b/=255;
@@ -1358,3 +1351,4 @@
     }
   }, 100);
 })();
+
