+++ package.json
@@ -1,0 +1,16 @@
+{
+  "name": "content-plan-tool",
+  "version": "1.0.0",
+  "description": "Dieses Repository enthält die Quelltexte des **Jahrescontent‑Kalenders**, eines webbasierten Werkzeugs zur Jahresplanung von Inhalten. Es ist vollständig clientseitig umgesetzt und speichert alle Daten im Browser‐Speicher (mit Fallback auf ein In‑Memory‑Store). Die Anwendung orientiert sich gestalterisch am modularen Drei‑Spalten‑Layout aus dem mitgelieferten Mockup.",
+  "main": "app.js",
+  "scripts": {
+    "test": "jest"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "type": "commonjs",
+  "devDependencies": {
+    "jest": "^30.0.5"
+  }
+}
