+++ todo.txt
@@ -1,5 +1,7 @@
 - [ ] Skip-Link in weiteren Seiten prüfen
-- [ ] Weitere Tooltips für Einstellungen ergänzen
-- [ ] Kontrastwerte im Dark-Theme testen
+- [x] Weitere Tooltips für Einstellungen ergänzen
+- [x] Kontrastwerte im Dark-Theme testen
 - [ ] Farbschema weiter an Vorlage anpassen
-- [ ] Module als Kartenlayout darstellen
+- [x] Module als Kartenlayout darstellen
+- [ ] Weitere Tests für Kalenderfunktionen ergänzen
+- [ ] Kartenbarrierefreiheit überprüfen
